requirements.txt
fastapi==0.110.1
sqlalchemy==2.0.29
asyncpg==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
uvicorn==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
